feature: entire project reestruction to supabaseauth
