+++ app/page.tsx
@@ -2,20 +2,11 @@ import Header from '../components/Header/Header'
 import Sidebar from '../components/Sidebar/Sidebar'
 import GetAuth from '../components/GetAuth/GetAuth'
 import { createClient } from '@/utils/supabase/server'
-import Sidobar from '../components/Sidobar/Sidobar'
+import ClaimedChannels from "@/components/ClaimedChannels/ClaimedChannel"
 
 export default async function Home() {
     const supabase = createClient()
-    const { data: notes } = await supabase.from('notes').select()
 
-    const fetchChannelInfo = async () => {
-        try {
-            console.log('teste')
-        } catch {
-            console.log('errooooooo')
-        }
-    }
-    fetchChannelInfo()
     return (
         <main className="flex min-h-screen flex-col">
             <Header />
@@ -24,8 +15,6 @@ export default async function Home() {
                     <Sidebar />
                     <div className="flex">
                         <p>home</p>
-                        <Sidobar />
-                        {/* <pre>{JSON.stringify(notes, null, 2)}</pre> */}
                     </div>
                 </div>
             </div>
